Add files via upload
-------------------------------------
Vviolet Hair Salon README
-------------------------------------


Overview:
Hello! This project was created by Ashlyn Emerson and Greer Pirtle. 
It was very intriguing to create a website for a business on our own. 
During the time we worked on the website, we got along very well and had fun while completing it!
Together we have gained a lot of new skills on building a website and the different properties involved in its creation.
Our website is a great demonstration of how we applied our skills and worked as a team.


Home Page:
We began by choosing a color scheme that would complement the logo of the hair salon. We also decided to give some information about Vviolet Hair Salon and provided visuals for the customer or reader to see the top-notch qualities of Vviolet Hair Salon.


About Page:
In our opinion, the about page was the easiest page to create. On this page we discussed the history of the business and its founder. 
We als